`zetteldeft--check` should compare true filenames on both sides.
The function `zetteldeft--check` should expand symlinks in the
`deft-directory` as well as the `buffer-file-name`. If the
`true-filename` isn't compared on both, you'll have false mismatches
when the `deft-directory` is in a symlinked folder, like `~/org/`
which is a symlink to `~/Dropbox/org`.

Before the change, this would search for `~/Dropbox/org/` in
`~/org/new-zettel.org`. Now it searches for `~/Dropbox/org/` in
`~/Dropbox/org/new-zettel.org`.