Enable optional insertion of title line when renaming note
Previously, the title line of a note was automatically updated when renaming the
note; however, if a note did not already have a title line, this was simply
ignored. This commit allows a title line to be inserted in the latter case. This
behavior is controlled by a customizable variable, so that prior behavior can
also be achieved.